% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_mml.R
\name{fetch_ogc_api_mml}
\alias{fetch_ogc_api_mml}
\title{A internal helper}
\usage{
fetch_ogc_api_mml(
  api_url,
  limitti,
  custom_params,
  mml_apikey = getOption("geofi_mml_api_key")
)
}
\arguments{
\item{api_url}{}
}
\description{
A internal helper
}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
