% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_statfi.R
\name{fetch_ogc_api_statfi}
\alias{fetch_ogc_api_statfi}
\title{A internal helper}
\usage{
fetch_ogc_api_statfi(api_url, limitti)
}
\arguments{
\item{api_url}{}
}
\description{
A internal helper
}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
