% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fin_4500_3067}
\alias{fin_4500_3067}
\title{A simple feature point data of Finnish land borders in 1:4500 resolution in EPSG:3067 coordinate system}
\format{
A simple feature MULTIPOLYGON data with 1 row and 3 variables:
\describe{
\item{name_fi{Name in Finnish}
\item{name_fen{Name in Swedish}
\item{name_en}{Name in English}
}
}
\usage{
fin_4500_3067
}
\description{
A simple feature point data of Finnish land borders in 1:4500 resolution in EPSG:3067 coordinate system
}
\keyword{datasets}
