\docType{package}
\name{fingis-package}
\alias{fingis}
\alias{fingis-package}
\title{Finland GIS tools for R}
\description{
Brief summary of the helsinki package
}
\details{
\tabular{ll}{ Package: \tab fingis\cr Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014\cr License: \tab FreeBSD\cr LazyLoad: \tab
yes\cr }

R Tools for helsinki API
}
\examples{
# library(fingis)
}
\author{
Juuso Parkkinen, Leo Lahti and Joona Lehtomaki
\email{louhos@googlegroups.com}
}
\references{
See citation("fingis")
\url{http://ropengov.github.com/fingis}
}
\keyword{package}

