% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode Finnish place names or street addresses using Finnish National Land Survey geocoding service
Thin wrapper around geocoding REST -api by \href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{National Land Survey} documented at
\href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{technical description geocoding service REST}.}
\usage{
geocode(
  search_string,
  source = "interpolated-road-addresses",
  output_crs = "EPSG:3067",
  lang = "fi",
  size = NULL,
  options = NULL,
  api_key = getOption("geofi_mml_api_key")
)
}
\arguments{
\item{search_string}{string place name or an address}

\item{source}{string Source of data. One of 'interpolated-road-addresses','geographic-names', 'addresses','mapsheets-tm35' and 'cadastral-units'}

\item{output_crs}{string CRS of the output sf object. One of 'EPSG:3067' or 'EPSG:4326'}

\item{lang}{string language for values in classification api fi, sv, en}

\item{size}{numeric Number of hits returned}

\item{options}{options}

\item{api_key}{MML api key}
}
\value{
sf object
}
\description{
Geocode Finnish place names or street addresses using Finnish National Land Survey geocoding service
Thin wrapper around geocoding REST -api by \href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{National Land Survey} documented at
\href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{technical description geocoding service REST}.
}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
