% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_reverse}
\alias{geocode_reverse}
\title{Reverse geocode geographic locations into Finnish place names or street addresses using Finnish National Land Survey geocoding service
Thin wrapper around geocoding REST -api by \href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{National Land Survey} documented at
\href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{technical description geocoding service REST}.}
\usage{
geocode_reverse(
  point,
  boundary_circle_radius = NULL,
  size = NULL,
  layers = NULL,
  sources = NULL,
  boundary_country = NULL,
  return = "sf",
  api_key = getOption("geofi_mml_api_key")
)
}
\arguments{
\item{point}{string location plane name or an address}

\item{boundary_circle_radius}{numeric Boundary}

\item{size}{numeric}

\item{layers}{string}

\item{sources}{string}

\item{boundary_country}{string}

\item{api_key}{string  MML api key}
}
\value{
sf object or list
}
\description{
Reverse geocode geographic locations into Finnish place names or street addresses using Finnish National Land Survey geocoding service
Thin wrapper around geocoding REST -api by \href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{National Land Survey} documented at
\href{https://www.maanmittauslaitos.fi/kartat-ja-paikkatieto/ammattilaiskayttajille/paikkatietojen-rajapintapalvelut/geokoodauspalvelu}{technical description geocoding service REST}.
}
\examples{
suomenlinna <- data.frame(lon = 24.933333, lat = 60.1725) |>
  (\(.) sf::st_as_sf(., coords = c("lon","lat"), crs = 4326))()
geocode_reverse(point = suomenlinna,
            source = "geographic-names")
}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
