% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{GeoStatFi-class}
\alias{GeoStatFi}
\alias{GeoStatFi-class}
\title{Retrieves geospatial data from stat.fi.}
\value{
In case the service at stat.fi cannot be reached, the relevant methods return \code{character(0)}.
}
\description{
Retrieves geospatial data from Statistics Finland (http://www.stat.fi).
For more information, see http://www.stat.fi/tup/rajapintapalvelut/index_en.html.
For General Terms of Use, see http://www.stat.fi/org/lainsaadanto/yleiset_kayttoehdot_en.html/.
}
\details{
Currently available data sets include:
\itemize{
  \item Population densities in various demographic groups
  \item Production and industrial facilities
  \item Educational institutions
  \item Road accidents
}
}
\section{Methods}{

\describe{
\item{\code{getEducationalInstitutions(layer = "oppilaitokset:oppilaitokset")}}{Retrieves educational institutions as a Spatial* object.}

\item{\code{getPopulation(layer)}}{Retrieves population grid data \code{layer} as a Spatial* object.}

\item{\code{getProductionAndIndustrialFacilities(layer = "ttlaitokset:toimipaikat")}}{Retrieves production and industrial facilities as a Spatial* object.}

\item{\code{getRaster(layer, template = raster(extent(85000, 726000, 6629000, 7777000),
  nrows = 1148, ncols = 641, crs =
  CRS("+proj=utm +zone=35 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")),
  fun = "count")}}{Converts SpatialPolygonsDataFrame object \code{layer} to Raster* object or rasterizes SpatialPointsDataFrame using the \code{template} RasterLayer object as a template with the function \code{fun}.}

\item{\code{getRoadAccidents(layer)}}{Retrieves road accident data \code{layer} as a Spatial* object.}

\item{\code{listPopulationLayers()}}{Returns a list of available population grid data sets as a character vector.}

\item{\code{listRoadAccidentsLayers()}}{Returns a list of available road accident data sets as a character vector.}
}}
\examples{
# See the vignette.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

