% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeoStatFi.R
\docType{data}
\name{GeoStatFiWFSRequest}
\alias{GeoStatFiWFSRequest}
\title{Builds a WFS request to the \code{stat.fi} geospatial API.}
\description{
Builds a WFS request to the \code{stat.fi} geospatial API. The request is submitted with
\code{\link{GeoStatFiWFSClient}} to retrieve the data.
For more information about the API, see \url{http://www.stat.fi/tup/rajapintapalvelut/index_en.html}.
For General Terms of Use, see \url{http://www.stat.fi/org/lainsaadanto/yleiset_kayttoehdot_en.html/}.
}
\section{Currently supported data sets}{

\itemize{
  \item Population densities in various demographic groups
  \item Production and industrial facilities
  \item Educational institutions
  \item Road accidents
}
}

\section{Methods}{

\itemize{
  \item\code{getPopulationLayers()}: A request for a list of available population grid data sets (layers).
  \item\code{getPopulation(layer)}: A request for population grid data \code{layer}.
  \item\code{getProductionAndIndustrialFacilities()}: A request for production and industrial facilities.
  \item\code{getEducationalInstitutions()}: A request for educational institutions.
  \item\code{getRoadAccidentsLayers()}: A request for a list of available road accident data sets as a character vector.
  \item\code{getRoadAccidents(layer)}: A request for road accident data \code{layer}.
  \item\code{getPostalCodeAreaLayers()}: A request for a list of available postal code area data sets as a character vector.
  \item\code{getPostalCodeArea(layer)}: A request for postal code area data \code{layer}.
}
}
\examples{
# See the vignette.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{GeoStatFiWFSClient}} \code{\link{WFSStreamingRequest}}
}
\keyword{datasets}

