% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buildings.R
\name{get_buildings}
\alias{get_buildings}
\title{Get geospatial data with all buildings and electoral districts from Väestörekisterikeskus}
\usage{
get_buildings()
}
\value{
sf-object
}
\description{
preprocessed geospatial data sf-objects
}
\examples{
 \dontrun{
 f <- get_buildings()
 plot(f)
 }

}
\author{
Markus Kainu <markus.kainu@kela.fi>
}
