% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{get_geocode}
\alias{get_geocode}
\title{Get geocode for given query}
\usage{
get_geocode(query, service = "okf", raw_query = F)
}
\arguments{
\item{query}{Either a street address, e.g. 'Mannerheimintie 100, Helsinki'
or place, e.g. 'Eduskuntatalo'}

\item{service}{Geocoding service to use, one of 'okf', 'openstreetmap' or 'google'}

\item{raw_query}{If true, don't prepend / append some default parameters to query.  Except for the ones specifying json format, send the query to API as-it-is}
}
\value{
A list with coordinates (lat, long) of the first output, and the raw output list
}
\description{
Get gecode for given query from one of the geocoding services:
OKF.fi Geocoding API Test Console: http://data.okf.fi/console/
OpenStreetMap Nominatim
(usage policy: \url{http://wiki.openstreetmap.org/wiki/Nominatim_usage_policy})
Google Maps API
See Terms and Conditions from \url{http://code.google.com/apis/maps/documentation/geocoding/}
}
\examples{
 \dontrun{
   # Geocode for a street address:
   gc1 <- get_geocode("Mannerheimintie 100, Helsinki");

   # Geocode for a place name
   gc2 <- get_geocode("&city=Helsinki", service="openstreetmap", raw_query=T)
}
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}, minor updates by Aaro Salosensaari \email{aaro.salosensaari@helsinki.fi} and Leo Lahti.
}
\references{
See citation("gisfin")
}

