\name{get_louhos_storage_path}
\alias{get_louhos_storage_path}
\title{Get Louhos storage path.}
\usage{
get_louhos_storage_path()
}
\value{
URL for Louhos data
}
\description{
Get url for Louhos storage. It is used to store some large
data files.
}
\examples{
url <- get_louhos_storage_path()
}
\author{
Leo Lahti and Juuso Parkkinen
\email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\keyword{utilities}

