% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/municipality.R
\name{get_municipality_map}
\alias{get_municipality_map}
\title{Get municipality border maps}
\usage{
get_municipality_map(data.source = "MML")
}
\arguments{
\item{data.source}{Specify the source of municipality map data: "MML" or "GADM"-}
}
\value{
\link{SpatialPolygonsDataFrame}-object in longlat-format.
}
\description{
Get municipality border maps
}
\details{
Retrieve municipality maps from
   MML (Land Survey Finland; Maanmittauslaitos, MML) (see \url{https://github.com/avoindata/mml}) or
   GADM \url{http://gadm.org/country}. For further maps from MML, see \code{\link{get_mml}}. For further maps from GADM, see \url{http://gadm.org/country}. Note that the GADM map data is not up-to-date.
}
\examples{
sp <- get_municipality_map("MML")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{get_mml}}
}

