% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_postalcode_areas}
\alias{get_postalcode_areas}
\title{Retrieve postal code areas as spatial data}
\usage{
get_postalcode_areas(data.dir = tempdir(), verbose = TRUE)
}
\arguments{
\item{data.dir}{A string. Specify a temporary folder for storing downloaded data.}

\item{verbose}{logical. Should R report extra information on progress?}
}
\value{
a spatial object (from SpatialPolygonsDataFrame class)
}
\description{
Retrieves postal code areas from http://www.palomaki.info/apps/pnro/
}
\details{
The data is created by Duukkis and licensed under CC BY 4.0
}
\examples{
\dontrun{pnro.sp <- get_postalcode_areas()}
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

