\name{get_theme_map}
\alias{get_theme_map}
\title{Get blank ggplot2 theme for plotting maps}
\usage{
get_theme_map()
}
\value{
theme_map A ggplot2 theme object
}
\description{
Get blank ggplot2 theme for plotting maps
}
\examples{
get_theme_map();
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}
\references{
See citation("fingis")
}

