% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs.R
\name{get_wfs_layer}
\alias{get_wfs_layer}
\title{Get geospatial data from a WFS}
\usage{
get_wfs_layer(url, layer, serviceVersion = "1.0.0", logger = "INFO")
}
\arguments{
\item{url}{Character source WFS URL}

\item{layer}{Character feature layer name}

\item{serviceVersion}{Character service version (default: "1.0.0")}

\item{logger}{Character logging level (default: "INFO")}
}
\value{
sf object
}
\description{
\code{get_wfs_layer()} retrieves geospatial data from a WFS source and turns the
retrieved data into a \code{sf} object.
}
\details{
Function assumes that the structure of WFS, including e.g. feature layer
name, is known by the user. In case you need to probe the WFS layer, have
a look at package \link{ows4R}.
}
\note{
In case the retrieved data does not have a coordinate reference
system (CRS) defined, \href{https://epsg.io/3067}{ETRS89 / TM35FIN} is assumed and
coerced.
}
\examples{
 \dontrun{
 url <- http://geo.stat.fi/geoserver/wfs
 layer <- paste0("tilastointialueet:kunta", 4500, "k_", 2019)
 version <- "1.0.0"
 sp_dat <- get_wfs_layer(url, layer, version)
 }

}
\author{
Markus Kainu \href{mailto:markus.kainu@kela.fi}{markus.kainu@kela.fi},
Joona Lehtomäki \href{mailto:joona.lehtomaki@iki.fi}{joona.lehtomaki@iki.fi}
}
