\name{GetGeocodeGoogleMaps}
\alias{GetGeocodeGoogleMaps}
\title{Get geo code from Google Map}
\usage{
GetGeocodeGoogleMaps(str)
}
\arguments{
  \item{str}{Street address, e.g. 'Mannerheimintie, 00100,
  FI'}
}
\value{
coordinates (lat, lon)
}
\description{
Get gecode for given street address from Google Maps API
See Terms and Conditions from
http://code.google.com/apis/maps/documentation/geocoding/
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}

