% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{gisfin-package}
\alias{gisfin}
\alias{gisfin-package}
\title{Finland GIS R tools}
\description{
Brief summary of the gisfin package
}
\details{
\tabular{ll}{
Package: \tab gisfin\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014\cr
License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr
}

R Tools for helsinki API
}
\examples{
# library(gisfin)
}
\author{
Juuso Parkkinen, Leo Lahti, Joona Lehtomaki and Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
\url{https://github.com/rOpenGov/gisfin}
}
\keyword{package}

