% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\docType{data}
\name{GMLFile}
\alias{GMLFile}
\title{A class for providing a file name reference to a GML document}
\description{
A class for providing a file name reference to a GML document.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

