% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_hyvinvointialue}
\alias{grid_hyvinvointialue}
\title{custom geofacet grid for Wellbeing services counties}
\format{
A data frame with 23 rows and 4 variables:
\describe{
\item{name}{Wellbeing services county name (hyvinvointialue) in Finnish}
\item{code}{Wellbeing services counties code}
\item{row}{Vertical location in grid}
\item{col}{Horizontal location in grid}
}
}
\usage{
grid_hyvinvointialue
}
\description{
Grid table to be used with ggplot2 and geofacet
}
\keyword{datasets}
