% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_mk_2020}
\alias{grid_mk_2020}
\title{custom geofacet grid for regions as in 2020}
\format{
A data frame with 19 rows and 4 variables:
\describe{
\item{name}{Region name (maakunta) in Finnish}
\item{code}{Region code}
\item{row}{Vertical location in grid}
\item{col}{Horizontal location in grid}
}
}
\usage{
grid_mk_2020
}
\description{
Grid table to be used with ggplot2 and geofacet
}
\keyword{datasets}
