% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_pohjois_karjala}
\alias{grid_pohjois_karjala}
\title{custom geofacet grid for Pohjois-Karjala region}
\format{
A data frame with 13 rows and 4 variables:
\describe{
\item{name}{Municipality name (kunta) in Finnish}
\item{code}{Municipality code}
\item{row}{Vertical location in grid}
\item{col}{Horizontal location in grid}
}
}
\usage{
grid_pohjois_karjala
}
\description{
Grid table to be used with ggplot2 and geofacet
}
\keyword{datasets}
