% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_sairaanhoitop_2019}
\alias{grid_sairaanhoitop_2019}
\title{custom geofacet grid for health care districts as in 2019}
\format{
A data frame with 21 rows and 4 variables:
\describe{
\item{name}{District name (Sairaanhoitopiiri) in Finnish}
\item{code}{District code}
\item{row}{Vertical location in grid}
\item{col}{Horizontal location in grid}
}
}
\usage{
grid_sairaanhoitop_2019
}
\description{
Grid table to be used with ggplot2 and geofacet
}
\keyword{datasets}
