% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_varsinais_suomi_2019}
\alias{grid_varsinais_suomi_2019}
\title{custom geofacet grid for Varsinais-Suomi region as in 2019}
\format{A data frame with 19 rows and 4 variables:
\describe{
\item{name}{Municipality name (kunta) in Finnish}
\item{code}{Municipality code}
\item{row}{Vertical location in grid}
\item{col}{Horizontal location in grid}
}}
\usage{
grid_varsinais_suomi_2019
}
\description{
Grid table to be used with ggplot2 and geofacet
}
\keyword{datasets}
