% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_codes_as_character.R
\name{integers_as_character}
\alias{integers_as_character}
\title{Convert an integer variable into character variable of varying width based on variable name.}
\usage{
integers_as_character(dat, varname)
}
\arguments{
\item{dat}{named vector}

\item{varname}{named vector}
}
\description{
Convert an integer variable into character variable of varying width based on variable name.
}
\note{
For internal use, not exported.
}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
