\name{list_mml_datasets}
\alias{list_mml_datasets}
\title{List Land Survey Finland (MML) data sets readily available in RData format.}
\usage{
  list_mml_datasets(...)
}
\arguments{
  \item{...}{Arguments to be passed

  Return:}
}
\value{
  List of data sets
}
\description{
  Arguments:
}
\details{
  Retrieves the list of currently available MML data sets
  in http://www.datavaalit.fi/storage/avoindata/mml/rdata/
}
\examples{
# list_mml_datasets()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

