\name{LouhosStoragePath}
\alias{LouhosStoragePath}
\title{LouhosStoragePath}
\usage{
LouhosStoragePath()
}
\value{
URL for Louhos data
}
\description{
Get url for Louhos storage.
}
\examples{
url <- LouhosStoragePath()
}
\author{
Leo Lahti and Juuso Parkkinen
\email{louhos@googlegroups.com}
}
\references{
See citation("fingis")
}
\keyword{utilities}

