\docType{data}
\name{MML}
\alias{MML}
\title{MML data documentation}
\format{list}
\usage{
sp <- get_MML(map.id = "4_5_milj_shape_etrs-tm35fin", data.id = "kunta4_p")
}
\description{
The MML data set contains GIS (geographic) data obtained
from Maanmittauslaitos (MML; Land Survey Finland).
Copyright (C) MML 2013. The open licenses allow
modification and redistribution of the data. The data sets
have been preprocessed to allow light-weight access through
R. For full details of the conversions, licenses, data
documentation etc., see
\url{https://github.com/louhos/takomo/tree/master/MML/Kapsi}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
\url{http://www.maanmittauslaitos.fi/aineistot-palvelut/latauspalvelut/avoimien-aineistojen-tiedostopalvelu}
}
\keyword{data}

