% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_central_localities}
\alias{municipality_central_localities}
\title{A simple feature point data containing locations of municipalities central localities}
\format{
A simple feature POINT data with 311 rows and 17 variables:
\describe{
\item{id}{id}
\item{mtk_id}{mtk_id}
\item{sijaintitarkkuus}{location precision}
\item{aineistolahde}{data source}
\item{alkupvm}{start date}
\item{teksti}{Municipality name}
\item{suunta}{direction}
\item{dx}{dx}
\item{dy}{dy}
\item{kohdeluokka}{kohdeluokka}
\item{ladontatunnus}{ladontatunnus}
\item{kirjasintyyppikoodi}{kirjasintyyppikoodi}
\item{kirjasinkoko}{kirjasinkoko}
\item{kirjasinvarikoodi}{kirjasinvarikoodi}
\item{kirjasinkallistuskulma}{kirjasinkallistuskulma}
\item{kirjasinvalyskerroin}{kirjasinvalyskerroin}
\item{kuntatunnus}{municipality code}
\item{geom}{geom}
}
}
\usage{
municipality_central_localities
}
\description{
Data is extracted from latest version (January 2021) of Topographic Database (Maastotietokanta) by National Land Survey of Finland (Maanmittauslaitos)
}
\keyword{datasets}
