% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_central_localities_df}
\alias{municipality_central_localities_df}
\title{A data frame containing locations of municipalities central localities}
\format{
A data.frame with 309 rows and 19 variables:
\describe{
\item{mtk_id}{mtk_id}
\item{sijaintitarkkuus}{location precision}
\item{aineistolahde}{data source}
\item{alkupvm}{start date}
\item{teksti}{Municipality name}
\item{suunta}{direction}
\item{dx}{dx}
\item{dy}{dy}
\item{kohderyhmä}{kohderyhmä}
\item{kohdeluokka}{kohdeluokka}
\item{ladontatunnus}{ladontatunnus}
\item{kirjasintyyppikoodi}{kirjasintyyppikoodi}
\item{kirjasinkoko}{kirjasinkoko}
\item{kirjasinvarikoodi}{kirjasinvarikoodi}
\item{kirjasinkallistuskulma}{kirjasinkallistuskulma}
\item{kirjasinvalyskerroin}{kirjasinvalyskerroin}
\item{kuntatunnus}{municipality code}
\item{X}{longitude}
\item{Y}{latitude}
}
}
\usage{
municipality_central_localities_df
}
\description{
Data is extracted from latest version (April 2025) of Topographic Database (Maastotietokanta) by National Land Survey of Finland (Maanmittauslaitos)
}
\keyword{datasets}
