% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2013}
\alias{municipality_key_2013}
\title{Vuoden 2013 kunnat erilaisilla alueluokituksilla}
\format{A data frame with 320 rows and 20 variables:
\describe{
  \item{kunta}{Kuntanumero}
  \item{name}{Kunnan nimi suomeksi}
  \item{avi_code}{Aluehallintoviraston koodi}
  \item{avi_nimi}{Aluehallintoviraston nimi}
  \item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen koodi}
  \item{ely_nimi}{Elinkeino-, liikenne- ja ympäristökeskuksen nimi}
  \item{kr_code}{Tilastokeskuksen tilastollisen kuntaluokituksen koodi}
  \item{kr_nimi}{Tilastokeskuksen tilastollisen kuntaluokituksen nimi}
  \item{mk_code}{Maakunnan koodi}
  \item{mk_nimi}{Maakunnan nimi}
  \item{sa_code}{Suuralueen koodi}
  \item{sa_nimi}{Suuralueen nimi}
  \item{sk_code}{Seutukunnan koodi}
  \item{sk_nimi}{Seutukunnan nimi}
  \item{sp_code}{Sairaanhoitopiirin koodi}
  \item{sp_nimi}{Sairaanhoitopiirin nimi}
  \item{tk_code}{Työssäköyntialueen koodi}
  \item{tk_nimi}{Työssäkäyntialueen nimi}
  \item{va_code}{Vaalipiirin koodi}
  \item{va_nimi}{Vaalipiirin nimi}
}}
\source{
\url{http://www.stat.fi/meta/luokitukset/kunta/001-2013/luokitusavaimet.html}
}
\usage{
municipality_key_2013
}
\description{
Data jolla kuntatason datoja voidaan aggregoida ylemmälle tasolla
}
\keyword{datasets}
