% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2014}
\alias{municipality_key_2014}
\title{Vuoden 2014 kunnat erilaisilla alueluokituksilla}
\format{A data frame with 320 rows and 14 variables:
\describe{
  \item{kunta}{Kuntanumero}
  \item{name}{Kunnan nimi suomeksi}
  \item{avi_code}{Aluehallintoviraston koodi}
  \item{avi_name}{Aluehallintoviraston nimi}
  \item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen koodi}
  \item{ely_name}{Elinkeino-, liikenne- ja ympäristökeskuksen nimi}
  \item{kr_code}{Tilastokeskuksen tilastollisen kuntaluokituksen koodi}
  \item{kr_name}{Tilastokeskuksen tilastollisen kuntaluokituksen nimi}
  \item{mk_code}{Maakunnan koodi}
  \item{mk_name}{Maakunnan nimi}
  \item{sa_code}{Suuralueen koodi}
  \item{sa_name}{Suuralueen nimi}
  \item{sp_code}{Sairaanhoitopiirin koodi}
  \item{sp_name}{Sairaanhoitopiirin nimi}
}}
\source{
\url{http://www.stat.fi/meta/luokitukset/kunta/001-2014/luokitusavaimet.html}
}
\usage{
municipality_key_2014
}
\description{
Data jolla kuntatason datoja voidaan aggregoida ylemmälle tasolla
}
\keyword{datasets}
