% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2017}
\alias{municipality_key_2017}
\title{Vuoden 2017 kunnat erilaisilla alueluokituksilla}
\format{A data frame with 311 rows and 28 variables:
\describe{
  \item{kunta}{Kuntanumero}
  \item{name}{Kunnan nimi suomeksi}
  \item{avi_code}{Aluehallintoviraston koodi}
  \item{avi_name}{Aluehallintoviraston nimi}
  \item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen koodi}
  \item{ely_name}{Elinkeino-, liikenne- ja ympäristökeskuksen nimi}
  \item{hp_code}{Hovioikeuspiirin koodi}
  \item{hp_name}{Hovioikeuspiirin nimi}
  \item{kr_code}{Tilastokeskuksen tilastollisen kuntaluokituksen koodi}
  \item{kr_name}{Tilastokeskuksen tilastollisen kuntaluokituksen nimi}
  \item{mk_code}{Maakunnan koodi}
  \item{mk_name}{Maakunnan nimi}
  \item{sa_code}{Suuralueen koodi}
  \item{sa_name}{Suuralueen nimi}
  \item{sk_code}{Seutukunnan koodi}
  \item{sk_name}{Seutukunnan nimi}
  \item{sp_code}{Sairaanhoitopiirin koodi}
  \item{sp_name}{Sairaanhoitopiirin nimi}
  \item{tk_code}{Työssäköyntialueen koodi}
  \item{tk_name}{Työssäkäyntialueen nimi}
  \item{va_code}{Vaalipiirin koodi}
  \item{va_name}{Vaalipiirin nimi}
  \item{vakuutuspiiri_name}{Kelan vakuutuspiirin nimi}
  \item{vakuutuspiiri_code}{Kelan vakuutuspiirin koodi (epävirallinen)}
  \item{asumistukialue_name}{Kelan asumistukialueen nimi}
  \item{asumistukialue_code}{Kelan asumistukialueen koodi (epävirallinen)}
  \item{erva_name}{Erityisvastuualueen nimi}
  \item{erva_code}{Erityisvastuualueen koodi (epävirallinen)}
  \item{avi_name_sv}{avi_name_sv}
  \item{ely_name_sv}{ely_name_sv}
  \item{kr_name_sv}{kr_name_sv}
  \item{mk_name_sv}{mk_name_sv}
  \item{sp_name_sv}{sp_name_sv}
  \item{vakuutuspiiri_name_sv}{vakuutuspiiri_name_sv}
}}
\source{
\url{http://www.stat.fi/meta/luokitukset/kunta/001-2017/luokitusavaimet.html}
}
\usage{
municipality_key_2017
}
\description{
Data jolla kuntatason datoja voidaan aggregoida ylemmälle tasolla
}
\keyword{datasets}
