% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2024}
\alias{municipality_key_2024}
\title{Municipality key table for 2024}
\format{
A data frame with 309 rows and 65 variables:
\describe{
\item{kunta}{Municipality code}
\item{municipality_name_fi}{Municipality name in Finnish}
\item{municipality_name_sv}{Municipality name in Swedish}
\item{municipality_name_en}{Municipality name in English}
\item{kuntaryhmitys_code}{Tilastollinen kuntaryhmitys (Statistical grouping) code}
\item{kuntaryhmitys_name_fi}{Tilastollinen kuntaryhmitys (Statistical grouping) name in Finnish}
\item{kuntaryhmitys_name_sv}{Tilastollinen kuntaryhmitys (Statistical grouping) name in Swedish}
\item{kuntaryhmitys_name_en}{Tilastollinen kuntaryhmitys (Statistical grouping) name in English}
\item{avi_code}{Aluehallintovirasto code (Regional State Administrative Agencies)}
\item{avi_name_fi}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{avi_name_sv}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{avi_name_en}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen code (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_fi}{Elinkeino-, liikenne- ja ympäristökeskuksen name in Finnish (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_sv}{Elinkeino-, liikenne- ja ympäristökeskuksen name in Swedish (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_en}{Elinkeino-, liikenne- ja ympäristökeskuksen name in English (Centre for Economic Development, Transport and the Environment)}
\item{maakunta_code}{Maakunta code (Regions of Finland)}
\item{maakunta_name_fi}{Maakunta name in Finnish (Regions of Finland)}
\item{maakunta_name_sv}{Maakunta name in Swedish (Regions of Finland)}
\item{maakunta_name_en}{Maakunta name in English (Regions of Finland)}
\item{seutukunta_code}{Seutukunta code (Sub-regions of Finland)}
\item{seutukunta_name_fi}{Seutukunta name in Finnish (Sub-regions of Finland)}
\item{seutukunta_name_sv}{Seutukunta name in Swedish (Sub-regions of Finland)}
\item{seutukunta_name_en}{Seutukunta name in English (Sub-regions of Finland)}
\item{year}{data year}
\item{suuralue_code}{Suuralue code}
\item{suuralue_name_fi}{Suuralueen nimi in Finnish (Large Areas)}
\item{suuralue_name_sv}{Suuralueen nimi in Swedish (Large Areas)}
\item{suuralue_name_en}{Suuralueen nimi in English (Large Areas)}
\item{nuts1_code}{NUTS 2016 classification code (level 1)}
\item{nuts1_name_fi}{NUTS 2016 classification name in Finnish (level 1)}
\item{nuts1_name_sv}{NUTS 2016 classification name in Swedish (level 1)}
\item{nuts1_name_en}{NUTS 2016 classification name in English (level 1)}
\item{nuts2_code}{NUTS 2016 classification code (level 2)}
\item{nuts2_name_fi}{NUTS 2016 classification name in Finnish (level 2)}
\item{nuts2_name_sv}{NUTS 2016 classification name in Swedish (level 2)}
\item{nuts2_name_en}{nuts2_name_en}
\item{nuts3_code}{NUTS 2016 classification code (level 2)}
\item{nuts3_name_fi}{NUTS 2016 classification name in Finnish (level 3)}
\item{nuts3_name_sv}{NUTS 2016 classification name in Swedish (level 3)}
\item{nuts3_name_en}{NUTS 2016 classification name in English (level 3)}
\item{vaalipiiri_code}{Vaalipiiri code (Constituencies)}
\item{vaalipiiri_name_fi}{Vaalipiiri name in Finnish (Constituencies)}
\item{vaalipiiri_name_sv}{Vaalipiiri name in Swedish (Constituencies)}
\item{vaalipiiri_name_en}{Vaalipiiri name in English (Constituencies)}
\item{municipality_code}{Municipality code}
\item{kunta_name}{Municipality name in Finnish}
\item{name_fi}{Municipality name in Finnish}
\item{name_sv}{Municipality name in Finnish}
\item{yhteistyoalue_code}{Sosiaali- ja terveydenhuollon yhteistyöalueiden koodi (Collaborative areas for healthcare and social welfare codes)}
\item{yhteistyoalue_name_fi}{Sosiaali- ja terveydenhuollon yhteistyöalueet}
\item{yhteistyoalue_name_sv}{Samarbetsområden för social- och hälsovården}
\item{yhteistyoalue_name_en}{Collaborative areas for healthcare and social welfare}
\item{kela_vakuutuspiiri_name_fi}{Kelan vakuutuspiiri name in Finnish (Insurance District)}
\item{kela_vakuutuspiiri_code}{Kelan vakuutuspiiri unofficial code (Insurance District)}
\item{kela_vakuutuspiiri_name_sv}{Kelan vakuutuspiiri name in Swedish (Insurance District)}
\item{kela_vakuutuspiiri_name_en}{Kelan vakuutuspiiri name in English (Insurance District)}
\item{kela_asumistukialue_name_fi}{Kelan asumistuen kuntaryhmät name in Finnish (Municipality categories for housing allowance)}
\item{kela_asumistukialue_code}{Kelan asumistuen kuntaryhmät unofficial code (Municipality categories for housing allowance)}
\item{kela_asumistukialue_name_sv}{Kelan asumistuen kuntaryhmät name in Swedish (Municipality categories for housing allowance)}
\item{kela_asumistukialue_name_en}{Kelan asumistuen kuntaryhmät name in English (Municipality categories for housing allowance)}
\item{hyvinvointialue_name_fi}{Hyvinvointialue name in Finnish (Wellbeing services counties)}
\item{hyvinvointialue_name_sv}{Hyvinvointialue name in Swedish (Wellbeing services counties)}
\item{hyvinvointialue_name_en}{Hyvinvointialue name in English (Wellbeing services counties)}
\item{hyvinvointialue_code}{Hyvinvointialue code (Wellbeing services counties)}
}
}
\source{
\url{https://data.stat.fi/api/classifications/v2/classifications}
}
\usage{
municipality_key_2024
}
\description{
Table for aggregating municipality level data to various regional groupings
}
\keyword{datasets}
