\name{natural.colors}
\alias{natural.colors}
\title{natural.colors: map colors}
\usage{
  natural.colors(...)
}
\arguments{
  \item{...}{Arguments to be passed.

  Returns:}
}
\value{
  colorlut
}
\description{
  Arguments:
}
\examples{
#
}
\author{
  Janne Aukia. Contact: \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

