\name{NormalizeValues}
\alias{NormalizeValues}
\title{NormalizeValues: normalize coordinate values (auxiliary function)}
\usage{
  NormalizeValues(mat)
}
\arguments{
  \item{mat}{data matrix

  Returns:}
}
\value{
  normalized data matrix
}
\description{
  Arguments:
}
\examples{
#
}
\author{
  Janne Aukia. Contact: \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

