% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_mml.R
\name{ogc_get_buildings_collections}
\alias{ogc_get_buildings_collections}
\title{Get all buildings related collections  \href{https://www.maanmittauslaitos.fi/rakennusten-kyselypalvelu/tekninen-kuvaus}{}}
\usage{
ogc_get_buildings_collections(api_key = getOption("geofi_mml_api_key"))
}
\arguments{
\item{api_key}{A string An \href{https://www.maanmittauslaitos.fi/en/rajapinnat/api-avaimen-ohje}{api key} is required to use NLS api services.}
}
\value{
data.frame
}
\description{
Thin wrapper around OGC API for \href{https://www.maanmittauslaitos.fi/rakennusten-kyselypalvelu/tekninen-kuvaus}{buildings data} (In Finnish) provided by
\href{https://www.maanmittauslaitos.fi/en}{National Land Survey of Finland} (NLS).
}
\examples{
ogc_get_buildings_collections()

}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
