% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_nls.R
\name{ogc_get_dem}
\alias{ogc_get_dem}
\title{Download and Cache Finnish DEM via OGC API Processes}
\usage{
ogc_get_dem(
  input,
  geometry_type = c("bbox", "polygon", "mapsheet"),
  resolution = c("2m", "10m"),
  file_format = "TIFF",
  overwrite = FALSE,
  api_key = getOption("geofi_mml_api_key")
)
}
\arguments{
\item{input}{A character string bbox ("minx,miny,maxx,maxy"), an \code{sf} object (for polygon), or a character vector (map sheet ID).}

\item{geometry_type}{One of: "bbox", "polygon", or "mapsheet".}

\item{resolution}{DEM resolution: "2m" or "10m".}

\item{file_format}{File format to request. Default is "TIFF".}

\item{overwrite}{If TRUE, always re-download and replace existing cached file.}

\item{api_key}{API key (default from \code{getOption("geofi_mml_api_key")}).}
}
\value{
Invisibly returns a \code{terra::SpatRaster} object (or file path if \code{terra} is unavailable).
}
\description{
Downloads and caches a DEM raster for a specified bounding box, polygon, or map sheet ID
from the National Land Survey of Finland using the OGC API Processes.
}
\examples{
\dontrun{
# example code
dem_file <- plot(ogc_get_dem("22.235421,60.446375,22.257050,60.453402",geometry_type ="bbox", resolution = "2m"))
plot(dem_file)
}
}
