% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_mml.R
\name{ogc_get_maastotietokanta_collections}
\alias{ogc_get_maastotietokanta_collections}
\title{Get all collections provided by Maastotietokanta (Topographic Database)}
\usage{
ogc_get_maastotietokanta_collections(api_key = getOption("geofi_mml_api_key"))
}
\arguments{
\item{api_key}{A string An \href{https://www.maanmittauslaitos.fi/en/rajapinnat/api-avaimen-ohje}{api key} is required to use NLS api services.}
}
\value{
data.frame
}
\description{
Thin wrapper around OGC API for \href{https://www.maanmittauslaitos.fi/en/maps-and-spatial-data/expert-users/product-descriptions/topographic-database}{Topographic database} provided by
\href{https://www.maanmittauslaitos.fi/en}{National Land Survey of Finland} (NLS).
}
\examples{
ogc_get_maastotietokanta_collections()

}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
