% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_mml.R
\name{ogc_get_nimisto}
\alias{ogc_get_nimisto}
\title{Query OCG API features service on \href{https://www.maanmittauslaitos.fi/en/maps-and-spatial-data/expert-users/product-descriptions/geographic-names}{Geographic names} (Nimistö)}
\usage{
ogc_get_nimisto(
  search_string = NULL,
  output_crs = 3067,
  limit = 10,
  bbox = NULL,
  custom_params = NULL,
  api_key = getOption("geofi_mml_api_key")
)
}
\arguments{
\item{search_string}{A string}

\item{bbox}{\code{1000}}

\item{api_key}{A string An \href{https://www.maanmittauslaitos.fi/en/rajapinnat/api-avaimen-ohje}{api key} is required to use NLS api services.}
}
\value{
sf object
}
\description{
Thin wrapper around OGC API for \href{https://www.maanmittauslaitos.fi/en/maps-and-spatial-data/expert-users/product-descriptions/geographic-names}{Geographic names} provided by
\href{https://www.maanmittauslaitos.fi/en}{National Land Survey of Finland} (NLS).
}
\examples{
ogc_get_nimisto(collection = "hautausmaa", output_crs = 3067)

}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
