\name{PlotSurface}
\alias{PlotSurface}
\title{PlotSurface: visualize 3D surface}
\usage{
PlotSurface(mat, colorlut = NULL)
}
\arguments{
  \item{mat}{data matrix describing the surface}

  \item{colorlut}{surface colors

  Returns:}
}
\value{
Used for side effects. Return: NULL
}
\description{
Arguments:
}
\examples{
#
}
\author{
Janne Aukia. Contact: \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

