% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Population1-class}
\alias{Population1}
\alias{Population1-class}
\title{Retrieves population density in Finland on a grid.}
\description{
Reference class for retrieving population density in Finland on a 1 km x 1 km grid provided by Statistics Finland.
For inherited methods, see the base class \code{\link{GeoStatFi}} help.
For more information, see http://www.stat.fi/tup/rajapintapalvelut/index_en.html.
For General Terms of Use, see http://www.stat.fi/org/lainsaadanto/yleiset_kayttoehdot_en.html/.
}
\details{
The data is currently available for the years 2005, 2010, 2011, 2012 and 2013 and contains the following fields:
\describe{
  \item{VAESTO}{Total population density (km^2).}
  \item{MIEHET}{Population density of males (km^2).}
  \item{NAISET}{Population density of females (km^2).}
  \item{IKA_0_14}{Population density of individuals aged <15 years.}
  \item{IKA_15_64}{Population density of individuals aged 15--64 years.}
  \item{IKA_65_}{Population density of individuals aged >64 years.}
}
Unoccupied grid cells are marked with value -1 as well as cells outside Finland. For the cells with <10 individuals,
only total population density (VAESTO) is provided.
}
\section{Methods}{

\describe{
\item{\code{query(years, saveMetadata = FALSE)}}{Retrieves population grid for the years specified in vector \code{years}. To retain metadata for each grid cell, set \code{saveMetadata=TRUE}.}
}}
\examples{
x <- Population1()
x$query(c(2012, 2013))
plot(x$getRaster("VAESTO", 2012))
plot(x$getRaster("VAESTO", 2013))
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{GeoStatFi}}
}

