\name{ReadASC}
\alias{ReadASC}
\title{ReadASC: read ASC file
Routines for 3D landscape visualization
contributed by Janne Aukia 2013}
\usage{
ReadASC(filename)
}
\arguments{
  \item{filename}{Input file name

  Returns:}
}
\value{
data matrix
}
\description{
Arguments:
}
\examples{
#
}
\author{
Janne Aukia. Contact: \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\seealso{
ReadXYZ
}
\keyword{utilities}

