% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_plot.R
\name{region_plot}
\alias{region_plot}
\title{Fast wrapper for visualizing regions on a map}
\usage{
region_plot(sp, color, region, palette = c("darkblue", "blue", "white", "red",
  "darkred"), by = 20, main = "")
}
\arguments{
\item{sp}{\link{SpatialPolygonsDataFrame} object}

\item{color}{Colour variable for filling}

\item{region}{Grouping variable (the desired regions)}

\item{palette}{Colour palette}

\item{by}{Interval for colors}

\item{main}{Plot title}
}
\value{
\code{\link{ggplot2}} object
}
\description{
Fast wrapper for visualizing regions on a map.
}
\examples{
sp <- get_municipality_map("MML"); region_plot(sp, color = "Kieli_ni1", region = "kuntakoodi") 
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

