% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MML.R
\name{ropengov_storage_path}
\alias{ropengov_storage_path}
\title{ropengov storage path}
\usage{
ropengov_storage_path()
}
\arguments{
\item{...}{Arguments to pass}
}
\value{
URL for Louhos data
}
\description{
Storage path for rOpenGov files.
}
\examples{
url <- ropengov_storage_path()
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

