% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sotkadata_population}
\alias{sotkadata_population}
\title{Municipality level population data from Sotkanet}
\format{
A data frame with 7107 rows and 3 variables:
\describe{
\item{municipality_code}{municipality_code}
\item{primary.value}{primary.value}
\item{year}{year}
}
}
\usage{
sotkadata_population
}
\description{
This dataset contains population data at municipality level pulled from THL (Sotkanet) from 2000 to 2022
}
\keyword{datasets}
