% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sotkadata_swedish_speaking_pop}
\alias{sotkadata_swedish_speaking_pop}
\title{Municipality level Swedish speaking population numbers from Sotkanet}
\format{
A data frame with 5761 rows and 3 variables:
\describe{
\item{municipality_code}{municipality_code}
\item{indicator.title.fi}{indicator.title.fi}
\item{primary.value}{primary.value}
}
}
\usage{
sotkadata_swedish_speaking_pop
}
\description{
This dataset contains Swedish speaking population figures at municipality level pulled from THL (Sotkanet) from 2000 to 2022
}
\keyword{datasets}
