% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statfi_zipcode_population}
\alias{statfi_zipcode_population}
\title{Zipcode level population data from Statistics Finland}
\format{
A data frame with 3027 rows and 2 variables:
\describe{
\item{posti_alue}{posti_alue}
\item{X2022}{X2022}
}
}
\usage{
statfi_zipcode_population
}
\description{
This dataset contains population for each zipcode in Finland. Data is downloaded from Statistics Finland
}
\keyword{datasets}
