% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sf.R
\name{to_sf}
\alias{to_sf}
\title{Transform a wfs_api object into a sf object.}
\usage{
to_sf(api_obj)
}
\arguments{
\item{api_obj}{wfs api object}
}
\value{
sf object
}
\description{
Statistics Finland WFS API response object's XML (GML) content is temporarily written on disk
and then immediately read back in into a sf object.
}
\note{
For internal use, not exported.
}
\author{
Joona Lehtomäki \href{mailto:joona.lehtomaki@iki.fi}{joona.lehtomaki@iki.fi}
}
