% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs.R
\name{wfs}
\alias{wfs}
\title{Get geospatial data at kunta, maakunta, etc leves from MML}
\usage{
wfs(url, layer, serviceVersion = "1.0.0", logger = "INFO")
}
\value{
sf-object
}
\description{
preprocessed geospatial data sf-objects
}
\examples{
 \dontrun{
 f <- get_geo(year=2016,level="kunta")
 plot(f)
 }

}
\author{
Markus Kainu <markus.kainu@kela.fi>
}
