% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\docType{data}
\name{WFSCachingRequest}
\alias{WFSCachingRequest}
\title{An abstract class for building a URL reference to a WFS with a caching}
\description{
The abstract method \code{getURL} must be overloaded in a subclass to provide a request URL to a WFS service.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

