% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{WFSClient}
\alias{WFSClient}
\title{WFS client abstract class}
\description{
WFS client abstract class
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSStreamingClient}}, \code{\link{WFSCachingClient}}, \code{\link{WFSRequest}}
}
\keyword{datasets}

