% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\docType{data}
\name{WFSRequest}
\alias{WFSRequest}
\title{An abstract class for referencing a WFS or a GML document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WFSRequest
}
\description{
This class should be inherited and the abstract method \code{getDataSource} overloaded
in a subclass to provide a reference.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSClient}}, \code{\link{GMLFile}}
}
\keyword{datasets}

