% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\docType{data}
\name{WFSStreamingRequest}
\alias{WFSStreamingRequest}
\title{WFSStreamingRequest abstract class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WFSStreamingRequest
}
\description{
An abstract class for building a URL reference to a WFS.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

