% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_access.R
\name{check_api_access}
\alias{check_api_access}
\title{Check access to http://geo.stat.fi/geoserver/wfs}
\usage{
check_api_access()
}
\value{
a logical.
}
\description{
Check if R has access to resources at http://geo.stat.fi/geoserver/wfs
}
\examples{
 \dontrun{
   check_api_access()
 }
}
\author{
Markus Kainu \email{markus.kainu@kapsi.fi}
}
