% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{GeoStatFi-class}
\alias{GeoStatFi}
\alias{GeoStatFi-class}
\title{Retrieves geospatial data from stat.fi.}
\description{
Retrieves geospatial data from Statistics Finland (http://www.stat.fi).
For more information, see http://www.stat.fi/tup/rajapintapalvelut/index_en.html.
For General Terms of Use, see http://www.stat.fi/org/lainsaadanto/yleiset_kayttoehdot_en.html/.
}
\details{
Currently available data sets include:
\itemize{
  \item Population densities in various demographic groups
  \item Production and industrial facilities
  \item Educational institutions
  \item Road accidents
}
}
\section{Methods}{

\describe{
\item{\code{getEducationalInstitutions(layer = "oppilaitokset:oppilaitokset")}}{Get educational institutions.}

\item{\code{getPopulation(layer)}}{Get population grid data set 'layer'.}

\item{\code{getProductionAndIndustrialFacilities(layer = "ttlaitokset:toimipaikat")}}{Get production and industrial facilities.}

\item{\code{getRaster(layer, template = raster(extent(85000, 726000, 6629000, 7777000),
  nrows = 1148, ncols = 641, crs =
  CRS("+proj=utm +zone=35 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")),
  fun = "count")}}{Converts SpatialPolygonsDataFrame object 'layer' to Raster* object or rasterizes SpatialPointsDataFrame onto 'template' raster with 'fun' function.}

\item{\code{getRoadAccidents(layer)}}{Get road accident data set 'layer'.}

\item{\code{listPopulationLayers()}}{List population grid data sets.}

\item{\code{listRoadAccidentsLayers()}}{List road accident data sets.}
}}
\examples{
See the vignette.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

