% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{GeoStatFi-class}
\alias{GeoStatFi}
\alias{GeoStatFi-class}
\title{Abstract reference class for grid statistics provided by Statistics Finland.}
\description{
Abstract reference class for grid statistics provided by Statistics Finland. See the currently implemented
classes for the available data queries: \code{\link{Population1}}.
}
\details{
For more information, see http://www.stat.fi/tup/rajapintapalvelut/index_en.html.
For General Terms of Use, see http://www.stat.fi/org/lainsaadanto/yleiset_kayttoehdot_en.html/.
}
\section{Methods}{

\describe{
\item{\code{extract(xy, feature, year)}}{Extract values of feature \code{feature} at coordinates \code{xy} of type Spatial* in year \code{year}. Does not check if coordinates are outside Finland. Returns a vector of the extracted values.}

\item{\code{getCRS()}}{Returns a CRS object.}

\item{\code{getRaster(feature, year)}}{Returns a raster layer object for feature \code{feature} in year \code{year}.}

\item{\code{load(fileName)}}{Load rasters into the object from \code{fileName}.}

\item{\code{save(fileName)}}{Save rasters to location \code{fileName} for later access, so no need to download the data again.}
}}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{Population1}}
}

