% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_geocode}
\alias{get_geocode}
\title{Get geocode for given query}
\usage{
get_geocode(query, service = "okf")
}
\arguments{
\item{query}{Either a street address, e.g. 'Mannerheimintie 100, Helsinki'
or place, e.g. 'Eduskuntatalo'}

\item{service}{Geocoding service to use, one of 'okf', 'openstreetmap' or 'google'}
}
\value{
A list with coordinates (lat, long) of the first output, and the raw output list
}
\description{
Get gecode for given query from one of the geocoding services:
OKF.fi Geocoding API Test Console: http://data.okf.fi/console/
OpenStreetMap Nominatim
(usage policy: http://wiki.openstreetmap.org/wiki/Nominatim_usage_policy)
Google Maps API
See Terms and Conditions from http://code.google.com/apis/maps/documentation/geocoding/
}
\examples{
gc <- get_geocode("Mannerheimintie 100, Helsinki");
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

