% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helsinki_spatial.R
\name{get_helsinki_aluejakokartat}
\alias{get_helsinki_aluejakokartat}
\title{Retrieve District Boundaries in Helsinki}
\usage{
get_helsinki_aluejakokartat(map.specifier = NULL, data.dir = tempdir(),
  verbose = TRUE)
}
\arguments{
\item{map.specifier}{A string. Specify the name of the
Helsinki District Boundary data set to retrieve.
Run 'get_helsinki_aluejakokartat()' to see available options.}

\item{data.dir}{A string. Specify a temporary folder for storing downloaded data.}

\item{verbose}{logical. Should R report extra information on progress?}
}
\value{
a spatial object (from SpatialPolygonsDataFrame class)
}
\description{
Retrieves District Boundaries data (Aluejakorajat, Aanestysalueet) from Helsinki Real Estate Department (Helsingin kaupungin kiinteistovirasto, HKK) through the HKK website \url{http://ptp.hel.fi/avoindata/index.html}

The data (C) 2011 Helsingin kaupunkimittausosasto.
}
\examples{
sp.suuralue <- get_helsinki_aluejakokartat(map.specifier="suuralue"); 
          spplot(sp.suuralue, zcol="Name");
}
\author{
Juuso Parkkinen, Joona Lehtomaki and Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

