% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_mml}
\alias{get_mml}
\title{MML (Land Survey Finland) data}
\usage{
get_mml(map.id, data.id, verbose = TRUE)
}
\arguments{
\item{map.id}{data ID. See details.}

\item{data.id}{data ID. See details.}

\item{verbose}{logical. Should R report extra information on progress?}
}
\value{
url connection
}
\description{
Retrieve data from Land Survey Finland (Maanmittauslaitos, MML). The data are processed into .RData files due to their big size. See the preprocessing details in https://github.com/avoindata/mml.
}
\details{
To browse for available RData, run list_mml_datasets() or see https://github.com/avoindata/mml/tree/master/rdata.
}
\examples{
datasets <- list_mml_datasets();
          map.id = names(datasets)[[1]];
          data.id <- datasets[[map.id]][[1]];
          sp <- get_mml(map.id, data.id)
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

