% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_grid.R
\name{get_population_grid}
\alias{get_population_grid}
\title{Get Statistical grid data polygons at two different resolution}
\usage{
get_population_grid(year = 2017, resolution = 5)
}
\arguments{
\item{year}{A numeric for year of the administrative borders. Years available are
2005 and 2010-2018.}

\item{resolution}{A numeric 1 (1km x 1km) or 5 (5km x 5km)}
}
\value{
sf object
}
\description{
Thin wrapper around Finnish zip code areas provided by
\href{http://www.stat.fi/org/avoindata/paikkatietoaineistot/tilastoruudukko_1km_en.html}{Statistic Finland}.
}
\examples{
 \dontrun{
 f <- get_population_grid(year=2017)
 plot(f)
 }

}
\author{
Markus Kainu \href{mailto:markus.kainu@kela.fi}{markus.kainu@kela.fi}, Joona Lehtomäki \href{mailto:joona.lehtomaki@iki.fi}{joona.lehtomaki@iki.fi}
}
