% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zipcodes.R
\name{get_zipcodes}
\alias{get_zipcodes}
\title{Get geospatial data at kunta, maakunta, etc leves from MML}
\usage{
get_zipcodes(year = 2017)
}
\arguments{
\item{year}{A numeric for year of the administerative borders. Available are 2006, 2010, 2011, 2012, 2014, 2015, 2016, 2017.}
}
\value{
sf-object
}
\description{
preprocessed geospatial data sf-objects
}
\examples{
 \dontrun{
 f <- get_zipcodes(year=2016,level="kunta")
 plot(f)
 }

}
\author{
Markus Kainu <markus.kainu@kela.fi>
}
