% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{GMLFile}
\alias{GMLFile}
\title{A class for providing a file name reference to a GML document}
\description{
A class for providing a file name reference to a GML document
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

