% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{ip_location}
\alias{ip_location}
\title{Get IP location}
\usage{
ip_location(ip)
}
\arguments{
\item{ip}{IP address as character}
}
\value{
Latitude and longitude as a numeric vector
}
\description{
Get geographic coordinates for a given IP-address from \url{http://www.datasciencetoolkit.org//ip2coordinates/}
}
\note{
Modified from original version by Kay Cichini
}
\examples{
\dontrun{ip_location("137.224.252.10")}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\keyword{utilities}

