% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MML.R
\name{list_mml_datasets}
\alias{list_mml_datasets}
\title{List Land Survey Finland (MML) data sets readily available in RData format.}
\usage{
list_mml_datasets(verbose = TRUE)
}
\arguments{
\item{verbose}{logical. Should R report extra information on progress?}
}
\value{
List of data sets
}
\description{
Retrieves the list of currently available MML data sets in http://www.datavaalit.fi/storage/avoindata/mml/rdata/
}
\examples{
datasets = list_mml_datasets()
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\keyword{utilities}

