% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_central_localities}
\alias{municipality_central_localities}
\title{A simple feature point data containing locations of municipalities central localities}
\format{
A simple feature POINT data with 311 rows and 21 variables:
\describe{
\item{mtk_id}{mtk_id}
\item{sijaintitarkkuus}{location precision}
\item{aineistolahde}{data source}
\item{alkupvm}{start date}
\item{teksti}{Municipality name}
\item{suunta}{direction}
\item{dx}{dx}
\item{dy}{dy}
\item{kohderyhmä}{kohderyhmä}
\item{kohdeluokka}{kohdeluokka}
\item{ladontatunnus}{ladontatunnus}
\item{kirjasintyyppikoodi}{kirjasintyyppikoodi}
\item{kirjasinkoko}{kirjasinkoko}
\item{kirjasinvarikoodi}{kirjasinvarikoodi}
\item{kirjasinkallistuskulma}{kirjasinkallistuskulma}
\item{kirjasinvalyskerroin}{kirjasinvalyskerroin}
\item{kuntatunnus}{municipality code}
\item{municipality_code}{municipality code}
\item{municipality_name_fi}{municipality name in Finnish}
\item{municipality_name_sv}{municipality name in Swedish}
\item{municipality_name_en}{municipality name in English}
\item{geometry}{geometry}
}
}
\usage{
municipality_central_localities
}
\description{
Data is extracted from latest version (April 2023) of Topographic Database (Maastotietokanta) by National Land Survey of Finland (Maanmittauslaitos)
}
\keyword{datasets}
