% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_central_localities.R
\name{municipality_central_localities}
\alias{municipality_central_localities}
\title{Get up-to-date municipality central locations}
\usage{
municipality_central_localities()
}
\value{
sf object
}
\description{
Convert municipality central locations from on-board data frame into a sf point data
}
\examples{
 \dontrun{
 f <- municipality_central_localities()
 plot(f)
 }

}
\author{
Markus Kainu \href{mailto:markus.kainu@kela.fi}{markus.kainu@kela.fi}
}
