% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key}
\alias{municipality_key}
\title{Aggregated municipality key table for years 2013-2021}
\format{
A data frame with 2519 rows and 74 variables:
\describe{
\item{kunta}{Municipality code}
\item{municipality_name_fi}{Municipality name in Finnish}
\item{municipality_name_sv}{Municipality name in Swedish}
\item{municipality_name_en}{Municipality name in English}
\item{kuntaryhmitys_code}{Tilastollinen kuntaryhmitys (Statistical grouping) code}
\item{kuntaryhmitys_name_fi}{Tilastollinen kuntaryhmitys (Statistical grouping) name in Finnish}
\item{kuntaryhmitys_name_sv}{Tilastollinen kuntaryhmitys (Statistical grouping) name in Swedish}
\item{kuntaryhmitys_name_en}{Tilastollinen kuntaryhmitys (Statistical grouping) name in English}
\item{avi_code}{Aluehallintovirasto code (Regional State Administrative Agencies)}
\item{avi_name_fi}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{avi_name_sv}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{avi_name_en}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen code (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_fi}{Elinkeino-, liikenne- ja ympäristökeskuksen name in Finnish (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_sv}{Elinkeino-, liikenne- ja ympäristökeskuksen name in Swedish (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_en}{Elinkeino-, liikenne- ja ympäristökeskuksen name in English (Centre for Economic Development, Transport and the Environment)}
\item{maakunta_code}{Maakunta code (Regions of Finland)}
\item{maakunta_name_fi}{Maakunta name in Finnish (Regions of Finland)}
\item{maakunta_name_sv}{Maakunta name in Swedish (Regions of Finland)}
\item{maakunta_name_en}{Maakunta name in English (Regions of Finland)}
\item{kielisuhde_code}{Kielisuhde code (Language distribution 2020)}
\item{kielisuhde_name_fi}{Kielisuhde name in Finnish (Language distribution 2020)}
\item{kielisuhde_name_sv}{Kielisuhde name in Swedish (Language distribution 2020)}
\item{kielisuhde_name_en}{Kielisuhde name in English (Language distribution 2020)}
\item{seutukunta_code}{Seutukunta code (Sub-regions of Finland)}
\item{seutukunta_name_fi}{Seutukunta name in Finnish (Sub-regions of Finland)}
\item{seutukunta_name_sv}{Seutukunta name in Swedish (Sub-regions of Finland)}
\item{seutukunta_name_en}{Seutukunta name in English (Sub-regions of Finland)}
\item{year}{data year}
\item{suuralue_code}{Suuralue code}
\item{suuralue_name_fi}{Suuralueen nimi in Finnish (Large Areas)}
\item{suuralue_name_sv}{Suuralueen nimi in Swedish (Large Areas)}
\item{suuralue_name_en}{Suuralueen nimi in English (Large Areas)}
\item{sairaanhoitop_code}{Sairaanhoitopiiri code (Health care districts)}
\item{sairaanhoitop_name_fi}{Sairaanhoitopiiri name in Finnish (Health care districts)}
\item{sairaanhoitop_name_sv}{Sairaanhoitopiiri name in Swedish (Health care districts)}
\item{sairaanhoitop_name_en}{Sairaanhoitopiiri name in English (Health care districts)}
\item{nuts1_code}{NUTS 2016 classification code (level 1)}
\item{nuts1_name_fi}{NUTS 2016 classification name in Finnish (level 1)}
\item{nuts1_name_sv}{NUTS 2016 classification name in Swedish (level 1)}
\item{nuts1_name_en}{NUTS 2016 classification name in English (level 1)}
\item{nuts2_code}{NUTS 2016 classification code (level 2)}
\item{nuts2_name_fi}{NUTS 2016 classification name in Finnish (level 2)}
\item{nuts2_name_sv}{NUTS 2016 classification name in Swedish (level 2)}
\item{nuts2_name_en}{nuts2_name_en}
\item{nuts3_code}{NUTS 2016 classification code (level 2)}
\item{nuts3_name_fi}{NUTS 2016 classification name in Finnish (level 3)}
\item{nuts3_name_sv}{NUTS 2016 classification name in Swedish (level 3)}
\item{nuts3_name_en}{NUTS 2016 classification name in English (level 3)}
\item{tyossakayntial_code}{Työssäkäyntialue code}
\item{tyossakayntial_name_fi}{Työssäkäyntialue name in Finnish}
\item{vaalipiiri_code}{Vaalipiiri code (Constituencies)}
\item{vaalipiiri_name_fi}{Vaalipiiri name in Finnish (Constituencies)}
\item{vaalipiiri_name_sv}{Vaalipiiri name in Swedish (Constituencies)}
\item{vaalipiiri_name_en}{Vaalipiiri name in English (Constituencies)}
\item{municipality_code}{Municipality code}
\item{kunta_name}{Municipality name in Finnish}
\item{name_fi}{Municipality name in Finnish}
\item{name_sv}{Municipality name in Finnish}
\item{erva_code}{Sairaanhoidon erityisvastuualueen code (Catchment areas for highly specialised medical care)}
\item{erva_name_fi}{Sairaanhoidon erityisvastuualueen name in Finnish (Catchment areas for highly specialised medical care)}
\item{erva_name_sv}{Sairaanhoidon erityisvastuualueen name in Swedish (Catchment areas for highly specialised medical care)}
\item{erva_name_en}{Sairaanhoidon erityisvastuualueen name in English (Catchment areas for highly specialised medical care)}
\item{kela_vakuutuspiiri_name_fi}{Kelan vakuutuspiiri name in Finnish (Insurance District)}
\item{kela_vakuutuspiiri_code}{Kelan vakuutuspiiri unofficial code (Insurance District)}
\item{kela_vakuutuspiiri_name_sv}{Kelan vakuutuspiiri name in Swedish (Insurance District)}
\item{kela_vakuutuspiiri_name_en}{Kelan vakuutuspiiri name in English (Insurance District)}
\item{kela_asumistukialue_name_fi}{Kelan asumistuen kuntaryhmät name in Finnish (Municipality categories for housing allowance)}
\item{kela_asumistukialue_code}{Kelan asumistuen kuntaryhmät unofficial code (Municipality categories for housing allowance)}
\item{kela_asumistukialue_name_sv}{Kelan asumistuen kuntaryhmät name in Swedish (Municipality categories for housing allowance)}
\item{kela_asumistukialue_name_en}{Kelan asumistuen kuntaryhmät name in English (Municipality categories for housing allowance)}
\item{hyvinvointialue_name_fi}{Hyvinvointialue name in Finnish (health and social services counties)}
\item{hyvinvointialue_name_sv}{Hyvinvointialue name in Swedish (health and social services counties)}
\item{hyvinvointialue_code}{Hyvinvointialue code (health and social services counties)}
}
}
\source{
\url{https://data.stat.fi/api/classifications/v2/classifications}
}
\usage{
municipality_key
}
\description{
Table for aggregating municipality level data to various regional groupings
}
\keyword{datasets}
