% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2013}
\alias{municipality_key_2013}
\title{Municipality key table for 2013}
\format{
A data frame with 320 rows and 33 variables:
\describe{
\item{kunta}{Municipality code}
\item{municipality_name_fi}{Municipality name in Finnish}
\item{municipality_name_sv}{Municipality name in Swedish}
\item{municipality_name_en}{Municipality name in English}
\item{kuntaryhmitys_code}{Tilastollinen kuntaryhmitys (Statistical grouping) code}
\item{kuntaryhmitys_name_fi}{Tilastollinen kuntaryhmitys (Statistical grouping) name in Finnish}
\item{kuntaryhmitys_name_sv}{Tilastollinen kuntaryhmitys (Statistical grouping) name in Swedish}
\item{kuntaryhmitys_name_en}{Tilastollinen kuntaryhmitys (Statistical grouping) name in English}
\item{avi_code}{Aluehallintovirasto code (Regional State Administrative Agencies)}
\item{avi_name_fi}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{avi_name_sv}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{avi_name_en}{Aluehallintovirasto name in Finnish (Regional State Administrative Agencies)}
\item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen code (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_fi}{Elinkeino-, liikenne- ja ympäristökeskuksen name in Finnish (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_sv}{Elinkeino-, liikenne- ja ympäristökeskuksen name in Swedish (Centre for Economic Development, Transport and the Environment)}
\item{ely_name_en}{Elinkeino-, liikenne- ja ympäristökeskuksen name in English (Centre for Economic Development, Transport and the Environment)}
\item{maakunta_code}{Maakunta code (Regions of Finland)}
\item{maakunta_name_fi}{Maakunta name in Finnish (Regions of Finland)}
\item{maakunta_name_sv}{Maakunta name in Swedish (Regions of Finland)}
\item{maakunta_name_en}{Maakunta name in English (Regions of Finland)}
\item{kielisuhde_code}{Kielisuhde code (Language distribution 2020)}
\item{kielisuhde_name_fi}{Kielisuhde name in Finnish (Language distribution 2020)}
\item{kielisuhde_name_sv}{Kielisuhde name in Swedish (Language distribution 2020)}
\item{kielisuhde_name_en}{Kielisuhde name in English (Language distribution 2020)}
\item{seutukunta_code}{Seutukunta code (Sub-regions of Finland)}
\item{seutukunta_name_fi}{Seutukunta name in Finnish (Sub-regions of Finland)}
\item{seutukunta_name_sv}{Seutukunta name in Swedish (Sub-regions of Finland)}
\item{seutukunta_name_en}{Seutukunta name in English (Sub-regions of Finland)}
\item{year}{data year}
\item{municipality_code}{Municipality code}
\item{kunta_name}{Municipality name in Finnish}
\item{name_fi}{Municipality name in Finnish}
\item{name_sv}{Municipality name in Finnish}
}
}
\source{
\url{https://data.stat.fi/api/classifications/v2/classifications}
}
\usage{
municipality_key_2013
}
\description{
Table for aggregating municipality level data to various regional groupings
}
\keyword{datasets}
