% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2015}
\alias{municipality_key_2015}
\title{Municipality key table for 2015}
\format{A data frame with 317 rows and 16 variables:
\describe{
  \item{kunta}{Municipality code or kuntanumero}
  \item{name}{Municipality name in Finnish}
  \item{avi_code}{Aluehallintovirasto code (Regional State Administrative Agencies)}
  \item{avi_name}{Aluehallintovirasto name (Regional State Administrative Agencies)}
  \item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen code (Centre for Economic Development, Transport and the Environment)}
  \item{ely_name}{Elinkeino-, liikenne- ja ympäristökeskuksen name (Centre for Economic Development, Transport and the Environment)}
  \item{kr_code}{Tilastokeskuksen tilastollisen kuntaluokituksen code}
  \item{kr_name}{Tilastokeskuksen tilastollisen kuntaluokituksen name}
  \item{mk_code}{Maakunnan code (Regions of Finland NUTS 3)}
  \item{mk_name}{Maakunnan name (Regions of Finland NUTS 3)}
  \item{sa_code}{Suuralueen code (Large areas NUTS 2)}
  \item{sa_name}{Suuralueen name (Large areas NUTS 2)}
  \item{sk_code}{Seutukunnan code (Sub-Regions of Finland NUTS 3)}
  \item{sk_name}{Seutukunnan name (Sub-Regions of Finland NUTS 3)}
  \item{va_code}{Vaalipiirin code (Electoral district)}
  \item{va_name}{Vaalipiirin name (Electoral district)}
}}
\source{
\url{http://www.stat.fi/meta/luokitukset/kunta/001-2015/luokitusavaimet.html}
}
\usage{
municipality_key_2015
}
\description{
Table for aggregating municipality level data to various regional groupings
}
\keyword{datasets}
