% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2016}
\alias{municipality_key_2016}
\title{Vuoden 2016 kunnat erilaisilla alueluokituksilla}
\format{A data frame with 313 rows and 48 variables:
\describe{
  \item{kunta}{Kuntanumero}
  \item{name}{Kunnan nimi suomeksi}
  \item{avi_code}{Aluehallintoviraston koodi}
  \item{avi_name}{Aluehallintoviraston nimi}
  \item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen koodi}
  \item{ely_name}{Elinkeino-, liikenne- ja ympäristökeskuksen nimi}
  \item{kr_code}{Tilastokeskuksen tilastollisen kuntaluokituksen koodi}
  \item{kr_name}{Tilastokeskuksen tilastollisen kuntaluokituksen nimi}
  \item{mk_code}{Maakunnan koodi}
  \item{mk_name}{Maakunnan nimi}
  \item{sa_code}{Suuralueen koodi}
  \item{sa_name}{Suuralueen nimi}
  \item{sk_code}{Seutukunnan koodi}
  \item{sk_name}{Seutukunnan nimi}
  \item{va_code}{Vaalipiirin koodi}
  \item{va_name}{Vaalipiirin nimi}
  \item{avi_name_sv}{avi_name_sv}
  \item{ely_name_sv}{ely_name_sv}
  \item{kr_name_sv}{kr_name_sv}
  \item{mk_name_sv}{mk_name_sv}
}}
\source{
\url{http://www.stat.fi}
}
\usage{
municipality_key_2016
}
\description{
Data jolla kuntatason datoja voidaan aggregoida ylemmälle tasolla
}
\keyword{datasets}
