% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2017}
\alias{municipality_key_2017}
\title{municipality_key_2017}
\format{
A data frame with 311 rows and 55 variables:
\describe{
\item{kunta}{kunta}
\item{municipality_name_fi}{municipality_name_fi}
\item{municipality_name_sv}{municipality_name_sv}
\item{municipality_name_en}{municipality_name_en}
\item{avi_code}{avi_code}
\item{avi_name_fi}{avi_name_fi}
\item{avi_name_sv}{avi_name_sv}
\item{avi_name_en}{avi_name_en}
\item{kuntaryhmitys_code}{kuntaryhmitys_code}
\item{kuntaryhmitys_name_fi}{kuntaryhmitys_name_fi}
\item{kuntaryhmitys_name_sv}{kuntaryhmitys_name_sv}
\item{kuntaryhmitys_name_en}{kuntaryhmitys_name_en}
\item{tyossakayntial_code}{tyossakayntial_code}
\item{tyossakayntial_name_fi}{tyossakayntial_name_fi}
\item{ely_code}{ely_code}
\item{ely_name_fi}{ely_name_fi}
\item{ely_name_sv}{ely_name_sv}
\item{ely_name_en}{ely_name_en}
\item{maakunta_code}{maakunta_code}
\item{maakunta_name_fi}{maakunta_name_fi}
\item{maakunta_name_sv}{maakunta_name_sv}
\item{maakunta_name_en}{maakunta_name_en}
\item{seutukunta_code}{seutukunta_code}
\item{seutukunta_name_fi}{seutukunta_name_fi}
\item{seutukunta_name_sv}{seutukunta_name_sv}
\item{seutukunta_name_en}{seutukunta_name_en}
\item{kielisuhde_code}{kielisuhde_code}
\item{kielisuhde_name_fi}{kielisuhde_name_fi}
\item{kielisuhde_name_sv}{kielisuhde_name_sv}
\item{kielisuhde_name_en}{kielisuhde_name_en}
\item{year}{year}
\item{suuralue_code}{suuralue_code}
\item{suuralue_name_fi}{suuralue_name_fi}
\item{suuralue_name_sv}{suuralue_name_sv}
\item{suuralue_name_en}{suuralue_name_en}
\item{sairaanhoitop_code}{sairaanhoitop_code}
\item{sairaanhoitop_name_fi}{sairaanhoitop_name_fi}
\item{sairaanhoitop_name_sv}{sairaanhoitop_name_sv}
\item{sairaanhoitop_name_en}{sairaanhoitop_name_en}
\item{nuts1_code}{nuts1_code}
\item{nuts1_name_fi}{nuts1_name_fi}
\item{nuts1_name_sv}{nuts1_name_sv}
\item{nuts1_name_en}{nuts1_name_en}
\item{nuts2_code}{nuts2_code}
\item{nuts2_name_fi}{nuts2_name_fi}
\item{nuts2_name_sv}{nuts2_name_sv}
\item{nuts2_name_en}{nuts2_name_en}
\item{nuts3_code}{nuts3_code}
\item{nuts3_name_fi}{nuts3_name_fi}
\item{nuts3_name_sv}{nuts3_name_sv}
\item{nuts3_name_en}{nuts3_name_en}
\item{municipality_code}{municipality_code}
\item{kunta_name}{kunta_name}
\item{name_fi}{name_fi}
\item{name_sv}{name_sv}
}
}
\source{
\url{https://data.stat.fi/api/classifications/v2/classifications}
}
\usage{
municipality_key_2017
}
\description{
Table for aggregating municipality level data to various regional groupings
}
\keyword{datasets}
