% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2019}
\alias{municipality_key_2019}
\title{Vuoden 2019 kunnat erilaisilla alueluokituksilla}
\format{A data frame with 311 rows and 65 variables:
\describe{
  \item{kunta}{kunta}
  \item{name}{name}
  \item{avi_code}{avi_code}
  \item{avi_name}{avi_name}
  \item{ely_code}{ely_code}
  \item{ely_name}{ely_name}
  \item{kr_code}{kr_code}
  \item{kr_name}{kr_name}
  \item{mk_code}{mk_code}
  \item{mk_name}{mk_name}
  \item{sa_code}{sa_code}
  \item{sa_name}{sa_name}
  \item{sk_code}{sk_code}
  \item{sk_name}{sk_name}
  \item{sp_code}{sp_code}
  \item{sp_name}{sp_name}
  \item{erva_name}{erva_name}
  \item{erva_code}{erva_code}
  \item{va_code}{va_code}
  \item{va_name}{va_name}
  \item{vakuutuspiiri_name}{vakuutuspiiri_name}
  \item{vakuutuspiiri_code}{vakuutuspiiri_code}
  \item{asumistukialue_name}{asumistukialue_name}
  \item{asumistukialue_code}{asumistukialue_code}
  \item{avi_name_sv}{avi_name_sv}
  \item{ely_name_sv}{ely_name_sv}
  \item{mk_name_sv}{mk_name_sv}
  \item{sa_name_sv}{sa_name_sv}
  \item{sa_name_sk}{sa_name_sk}
  \item{sp_name_sv}{sp_name_sv}
  \item{va_name_sv}{va_name_sv}
  \item{vakuutuspiiri_name_sv}{vakuutuspiiri_name_sv}
}}
\source{
\url{http://www.stat.fi/meta/luokitukset/kunta/001-2019/luokitusavaimet.html}
}
\usage{
municipality_key_2019
}
\description{
Data jolla kuntatason datoja voidaan aggregoida ylemmälle tasolla
}
\keyword{datasets}
