% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_key_2019}
\alias{municipality_key_2019}
\title{Municipality key table for 2019}
\format{A data frame with 311 rows and 34 variables:
\describe{
  \item{kunta}{Municipality code}
  \item{kunta_name}{Municipality name in Finnish}
  \item{avi_code}{Aluehallintovirasto code (Regional State Administrative Agencies)}
  \item{avi_name}{Aluehallintovirasto name (Regional State Administrative Agencies)}
  \item{ely_code}{Elinkeino-, liikenne- ja ympäristökeskuksen code (Centre for Economic Development, Transport and the Environment)}
  \item{ely_name}{Elinkeino-, liikenne- ja ympäristökeskuksen name (Centre for Economic Development, Transport and the Environment)}
  \item{kr_code}{Tilastokeskuksen tilastollisen kuntaluokituksen code}
  \item{kr_name}{Tilastokeskuksen tilastollisen kuntaluokituksen name}
  \item{mk_code}{Maakunnan code (Regions of Finland NUTS 3)}
  \item{mk_name}{Maakunnan name (Regions of Finland NUTS 3)}
  \item{sa_code}{Suuralueen code (Large areas NUTS 2)}
  \item{sa_name}{Suuralueen name (Large areas NUTS 2)}
  \item{sk_code}{Seutukunnan code (Sub-Regions of Finland NUTS 3)}
  \item{sk_name}{Seutukunnan name (Sub-Regions of Finland NUTS 3)}
  \item{sp_code}{Sairaanhoitopiirin code (Health care regions)}
  \item{sp_name}{Sairaanhoitopiirin name (Health care regions)}
  \item{erva_name}{Erityisvastuualueen name}
  \item{erva_code}{Erityisvastuualueen koodi (epävirallinen)}
  \item{va_code}{Vaalipiirin code (Electoral district)}
  \item{va_name}{Vaalipiirin name (Electoral district)}
  \item{kela_vakuutuspiiri_name}{Kelan vakuutuspiirin name}
  \item{kela_vakuutuspiiri_code}{Kelan vakuutuspiirin koodi (epävirallinen)}
  \item{kela_asumistukialue_name}{Kelan asumistukialueen name}
  \item{kela_asumistukialue_code}{Kelan asumistukialueen koodi (epävirallinen)}
  \item{avi_name_sv}{Aluehallintovirasto name (Regional State Administrative Agencies) in Swedish}
  \item{ely_name_sv}{Elinkeino-, liikenne- ja ympäristökeskuksen name (Centre for Economic Development, Transport and the Environment) in Swedish}
  \item{mk_name_sv}{Maakunnan name (Regions of Finland NUTS 3) in Swedish}
  \item{sa_name_sv}{Suuralueen name (Large areas NUTS 2) in Swedish}
  \item{sa_name_sk}{sa_name_sk}
  \item{sp_name_sv}{Sairaanhoitopiirin name (Health care regions) in Swedish}
  \item{va_name_sv}{Vaalipiirin name (Electoral district) in Swedish}
  \item{kela_vakuutuspiiri_name_sv}{vakuutuspiiri_name_sv}
  \item{name_sv}{municipality name in Swedish}
  \item{name_fi}{municipality name in Finnish}
}}
\source{
\url{http://www.stat.fi/meta/luokitukset/kunta/001-2019/luokitusavaimet.html}
}
\usage{
municipality_key_2019
}
\description{
Table for aggregating municipality level data to various regional groupings
}
\keyword{datasets}
