% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_mml.R
\name{ogc_get_maastotietokanta}
\alias{ogc_get_maastotietokanta}
\title{Get a selected collection from \href{https://www.maanmittauslaitos.fi/en/maps-and-spatial-data/expert-users/product-descriptions/topographic-database}{Topographic database}}
\usage{
ogc_get_maastotietokanta(
  collection = "hautausmaa",
  output_crs = 3067,
  limit = NULL,
  bbox = NULL,
  custom_params = NULL,
  api_key = getOption("geofi_mml_api_key")
)
}
\arguments{
\item{collection}{A string}

\item{api_key}{A string An \href{https://www.maanmittauslaitos.fi/en/rajapinnat/api-avaimen-ohje}{api key} is required to use NLS api services.}
}
\value{
sf object
}
\description{
Thin wrapper around OGC API for \href{https://www.maanmittauslaitos.fi/en/maps-and-spatial-data/expert-users/product-descriptions/topographic-database}{Topographic database} provided by
\href{https://www.maanmittauslaitos.fi/en}{National Land Survey of Finland} (NLS).
}
\examples{
ogc_get_maastotietokanta(collection = "hautausmaa", output_crs = 3067)

}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
