% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_api_statfi.R
\name{ogc_get_statfi_area}
\alias{ogc_get_statfi_area}
\title{Get Finnish municipality (multi)polygons for different years and/or scales.}
\usage{
ogc_get_statfi_area(
  year = 2022,
  scale = 4500,
  tessellation = NULL,
  output_crs = 3067,
  limit = NULL,
  bbox = NULL
)
}
\arguments{
\item{year}{A numeric for year of the administrative borders. Available are
2020, 2021 and 2022.}

\item{scale}{A scale or resolution of the shape. Two options: \code{1000}
equals 1:1 000 000 and \code{4500} equals 1:4 500 000.}

\item{tessellation}{A scale or resolution of the shape. Two options: \code{1000}
equals 1:1 000 000 and \code{4500} equals 1:4 500 000.}
}
\value{
sf object
}
\description{
Thin wrapper around Finnish statistical area units OGC API by
\href{https://www.stat.fi/org/avoindata/paikkatietoaineistot/kuntapohjaiset_tilastointialueet_en.html}{Statistics Finland}.
}
\examples{
ogc_get_statfi_area(year = 2020,
            tessellation = "maakunta",
            output_crs = 4326,
            limit = 10)

}
\author{
Markus Kainu \href{mailto:markus.kainu@kapsi.fi}{markus.kainu@kapsi.fi}
}
