% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ropengov_storage_path}
\alias{ropengov_storage_path}
\title{ropengov_storage_path}
\usage{
ropengov_storage_path()
}
\value{
URL for Louhos data
}
\description{
Arguments:
  ... Arguments to pass
}
\details{
Return:
}
\examples{
url <- ropengov_storage_path()
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

