\name{sp2df}
\alias{sp2df}
\title{Transform data from sp to data frame for ggplot2}
\usage{
sp2df(sp, region)
}
\arguments{
  \item{sp}{A sp object to be transformed}

  \item{region}{A string specifying the region of interest}
}
\value{
A ggplot2 theme object
}
\description{
Transform data from sp to data frame for ggplot2
}
\examples{
# sp2df(sp, region.name);
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}
\references{
See citation("fingis")
}

