% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sp2df}
\alias{sp2df}
\title{Transform data from sp to data frame for ggplot2}
\usage{
sp2df(sp, region = NULL, verbose = TRUE)
}
\arguments{
\item{sp}{A spatial object to be transformed}

\item{region}{A string specifying the region of interest}

\item{verbose}{logical. Should R report extra information on progress?}
}
\value{
A ggplot2 theme object
}
\description{
Transform data from sp to data frame for ggplot2
}
\examples{
sp.suuralue <- get_helsinki_aluejakokartat(map.specifier="suuralue");
          # Need to load rgeos and maptools for ggplot2::fortify in sp2df()
          library(rgeos);
          library(maptools);
          # Transform to df and plot with ggplot2
          df.suuralue <- sp2df(sp.suuralue, "Name");
          library(ggplot2);
          theme_set(get_theme_map());
          ggplot(df.suuralue, aes(x=long, y=lat, fill=Name)) +
          geom_polygon() + theme(legend.position="none")
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}

