% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sf.R
\name{to_sf}
\alias{to_sf}
\title{Transform a fmi_api object into a sf object.}
\usage{
to_sf(api_obj)
}
\arguments{
\item{api_obj}{wfs api object}
}
\value{
sf object
}
\description{
FMI API response object's XML (GML) content is temporarily wrtitten on disk
and then immediately read back in into a sf object.
}
\note{
For internal use, not exported.
}
\examples{
  \dontrun{
    response <- fmi_api(request = "getFeature",
                        storedquery_id = "fmi::observations::weather::daily::timevaluepair",
                        starttime = "2019-01-01", endtime = "2019-01-04",
                        fmisid = 100946)
    sf_obj <- to_sf(response)
  }

}
\author{
Joona Lehtomäki \href{mailto:joona.lehtomaki@iki.fi}{joona.lehtomaki@iki.fi}
}
