% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WFSClient.R
\docType{data}
\name{WFSStreamingClient}
\alias{WFSStreamingClient}
\title{Streams response from a WFS}
\description{
Dispatches a WFS request and parses response from the stream directly.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSRequest}}, \code{\link{WFSCachingClient}}
}
\keyword{datasets}

