% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WFSRequest.R
\docType{data}
\name{WFSStreamingRequest}
\alias{WFSStreamingRequest}
\title{An abstract class for building a URL reference to a WFS}
\description{
An abstract class for building a URL reference to a WFS
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

